Add test_file.txt
I created this text file to test what Git would do if I modified a file
after it had already been staged, but not committed. This test shows
that modifying a previously staged file will create a situation where
the file is both 'staged' and 'not staged' simultaneously. Committing
at this point will commit the version of the file before the second round
of changes. To include the most recent changes to the file, you must
add the file for staging once again.